Python Output
This image took multiple hours to create. That is the average color of every single frame from 'Fight Club (1999)' written in lines side by side. This probably took >10 hours to make. The 'Fast image from video.py' will get 1 frame pers second. If the video is 60 FPS it will grab 1 frame of those 60 and then move on to the next frame in the next second. This image was created from "get output image from video.py" which reads every frame of the video. Looks prettier tbh, but took painfully long to make. The txt file is parsed fast enough in a matter of seconds. I believe 200k lines where paresed in 6-8 seconds.