CPP Output
This image was created in ~25 minutes. The text file is what takes the most time to make, so I used C++ to handle that, as it is a lot faster than Python and OpenCV (the library used to read the video frames), also has a library for C++. The file here also reads 1 frame per second, so if the video is running at 24fps, then the program will grab 1 frame, every 24 frames, skipping a lot of them that probably have the same average values.